move dialog out of the menu
