+++ app/routes/_app.playlists.view.$playlistId/components/playlist-menubar.tsx
@@ -1,27 +1,15 @@
 import {
 	Button,
-	Checkbox,
-	Dialog,
-	DialogContent,
-	DialogFooter,
-	DialogTrigger,
 	Menubar,
 	MenubarContent,
 	MenubarItem,
 	MenubarMenu,
 	MenubarTrigger,
-	ScrollArea,
-	Table,
-	TableBody,
-	TableCell,
-	TableHead,
-	TableHeader,
-	TableRow,
 } from "@/components";
 import { Game } from "@/types";
 import { PlusCircledIcon } from "@radix-ui/react-icons";
 import { useFetcher } from "@remix-run/react";
-import { useState } from "react";
+import { usePlaylistViewStore } from "@/store/playlist-view";
 
 interface PlaylistMenubarProps {
 	isPrivate: boolean;
@@ -37,16 +25,11 @@ interface PlaylistMenubarProps {
 
 export function PlaylistMenubar({
 	isPrivate,
-	userCollection,
-	playlistGames,
 	playlistId,
-	userId,
-	setRenameDialogOpen,
 	setDeletePlaylistDialogOpen,
-	isEditing,
-	setIsEditing,
 }: PlaylistMenubarProps) {
 	const markAsPrivateFetcher = useFetcher();
+	const store = usePlaylistViewStore();
 
 	const handleTogglePrivate = () => {
 		markAsPrivateFetcher.submit(
@@ -64,7 +47,9 @@ export function PlaylistMenubar({
 						<MenubarItem onClick={() => setDeletePlaylistDialogOpen(true)}>
 							Delete
 						</MenubarItem>
-						<MenubarItem onClick={() => setRenameDialogOpen(true)}>Rename</MenubarItem>
+						<MenubarItem onClick={() => store.setRenameDialogOpen(true)}>
+							Rename
+						</MenubarItem>
 						<MenubarItem onClick={handleTogglePrivate}>
 							{isPrivate ? "Make Public" : "Set as Private"}
 						</MenubarItem>
@@ -72,98 +57,19 @@ export function PlaylistMenubar({
 				</MenubarMenu>
 			</Menubar>
 			<Button
-				variant={isEditing ? "default" : "outline"}
-				onClick={() => setIsEditing(!isEditing)}
+				variant={store.isEditing ? "default" : "outline"}
+				onClick={() => store.setIsEditing(!store.isEditing)}
 			>
 				Edit
 			</Button>
-			<AddGameToPlaylistDialog
-				userCollection={userCollection}
-				playlistGames={playlistGames}
-				userId={userId}
-				playlistId={playlistId}
-			/>
+			<Button
+				variant={"outline"}
+				aria-label="Add games button"
+				onClick={() => store.setAddGameDialogOpen(true)}
+			>
+				<PlusCircledIcon className="mr-3" aria-hidden="true" />
+				<span>Add Games</span>
+			</Button>
 		</div>
-	);
-}
-
-interface AddGameToPlaylistDialogProps {
-	userCollection: Game[];
-	playlistGames: number[];
-	userId: string;
-	playlistId: string;
-}
-
-function AddGameToPlaylistDialog({
-	userCollection,
-	playlistGames,
-	userId,
-	playlistId,
-}: AddGameToPlaylistDialogProps) {
-	const addGameFetcher = useFetcher();
-	const [open, setOpen] = useState<boolean>(false);
-	return (
-		<Dialog aria-label="Add game to playlist dialog" open={open} onOpenChange={setOpen}>
-			<DialogTrigger asChild>
-				<Button variant={"outline"} aria-label="Add games button">
-					<PlusCircledIcon className="mr-3" aria-hidden="true" />
-					<span>Add Games</span>
-				</Button>
-			</DialogTrigger>
-			<DialogContent>
-				<addGameFetcher.Form
-					method="POST"
-					action={`/api/playlists/${playlistId}/games`}
-					aria-labelledby="addGameForm"
-					onSubmit={() => setOpen(false)}
-				>
-					<input type="hidden" name="addedBy" value={userId} />
-					<ScrollArea className="w-full h-[50vh]">
-						<Table aria-label="Games list">
-							<TableHeader>
-								<TableRow>
-									<TableHead>Title</TableHead>
-									<TableHead>Rating</TableHead>
-									<TableHead>Select</TableHead>
-								</TableRow>
-							</TableHeader>
-							<TableBody>
-								{userCollection.map((game) => (
-									<AddGameTableRow
-										key={game.id}
-										game={game}
-										inPlaylist={playlistGames.includes(game.gameId)}
-									/>
-								))}
-							</TableBody>
-						</Table>
-					</ScrollArea>
-					<DialogFooter>
-						<Button aria-label="Submit button" type="submit">
-							Add
-						</Button>
-					</DialogFooter>
-				</addGameFetcher.Form>
-			</DialogContent>
-		</Dialog>
-	);
-}
-
-function AddGameTableRow({ game, inPlaylist }: { game: Game; inPlaylist: boolean }) {
-	const [checked, setChecked] = useState(inPlaylist);
-	return (
-		<TableRow key={game.id}>
-			<TableCell onClick={() => setChecked(!checked)}>{game.title}</TableCell>
-			<TableCell>{game.rating ?? 0}</TableCell>
-			<TableCell className="flex items-center">
-				<Checkbox
-					value={game.gameId}
-					name="gameIds"
-					aria-label={`Select game ${game.title}`}
-					checked={checked}
-					onCheckedChange={() => setChecked(!checked)}
-				/>
-			</TableCell>
-		</TableRow>
 	);
 }
